+++ src/main/java/com/mkyong/services/TopoService.java
@@ -1,10 +1,7 @@
 package com.mkyong.services;
 
 
-import com.mkyong.entity.Secteur;
-import com.mkyong.entity.Site;
-import com.mkyong.entity.Topo;
-import com.mkyong.entity.User;
+import com.mkyong.entity.*;
 import com.mkyong.exception.RecordNotFoundException;
 import com.mkyong.repository.TopoRepository;
 import org.slf4j.Logger;
@@ -60,20 +57,38 @@ public class TopoService {
     /*Methode pour modifier une entité topo*/
     public Topo UpdateTopo(Topo entity) throws RecordNotFoundException {
 
-            Topo topoAModifier = getTopoById(entity.getIdTopo());
-            if(topoAModifier!=null) {
-                logger.info(" l'entité topo à modifier a été trouvée et modifiée");
-                topoAModifier.setNomTopo(entity.getNomTopo());
-                topoAModifier.setDescription(entity.getDescription());
-                topoAModifier.setDisponible(entity.getDisponible());
-                topoAModifier.setSites(entity.getSites());
-                topoAModifier.setLocation(entity.getLocation());
-                entity = topoRepository.save(topoAModifier);
-                logger.info(" retour de la nouvelle entité topo de createOrUpdateTopo qui a été sauvegardée et le topo est existant");
-                return entity;
-            } else {
-                throw new RecordNotFoundException("No user record exist for given id and to modify it");
+        Topo topoAModifier = getTopoById(entity.getIdTopo());
+        if(topoAModifier!=null) {
+            logger.info(" l'entité topo à modifier a été trouvée et modifiée");
+            topoAModifier.setNomTopo(entity.getNomTopo());
+            topoAModifier.setDescription(entity.getDescription());
+            topoAModifier.setDisponible(entity.getDisponible());
+
+            logger.info(" la date de parution est: "+ topoAModifier.toStringDateParution());
+
+            topoAModifier.setSites(entity.getSites());
+            // enregistrement du topo dans chaque site concerné
+            //List<Site> sitesModifies=new ArrayList<Site>();
+            for(int i=0;i<(topoAModifier.getSites()).size();i++){
+                List<Site> listeSites=new ArrayList<Site>();
+                listeSites.addAll(topoAModifier.getSites());
+                Site site=listeSites.get(i);
+                site.setTopo(entity);
+                Site siteModifie=siteService.UpdateSite(site);
+                //sitesModifies.add(siteModifie);
             }
+            //topoAModifier.setSites(sitesModifies);
+
+            topoAModifier.setLocation(entity.getLocation());
+
+            topoAModifier.setReservation((entity.getReservation()));
+
+            topoAModifier = topoRepository.save(topoAModifier);
+            logger.info(" retour de la nouvelle entité topo de UpdateTopo qui a été sauvegardée et le topo est existant");
+            return topoAModifier;
+        } else {
+            throw new RecordNotFoundException("No user record exist for given id and to modify it");
+        }
     }
 
     /*Methode pour creer une entité topo*/
@@ -83,24 +98,15 @@ public class TopoService {
         newTopo.setNomTopo(entity.getNomTopo());
         newTopo.setDescription(entity.getDescription());
         newTopo.setDateParution(today);
-        newTopo.setDisponible(entity.getDisponible());
+        newTopo.setDisponible(true);
         newTopo.setLocation(entity.getLocation());
-        newTopo.setOwner(user);
-        newTopo.setImage(entity.getImage());
-        newTopo.setSites(entity.getSites());
+        newTopo.setReservation(null);
 
         // enregistrement de l'image dans le topo
-        //imageService.stockerImage(newTopo.getImage(), user);
+        newTopo.setImage(entity.getImage());
+        imageService.stockerImage(newTopo.getImage(),user);
 
-       // logger.info(" les images de topo: "+ entity.getImage().getId());
-
-        // enregistrement du topo dans liste des topos de user
-        Collection listeTopos = user.getTopos();
-        listeTopos.add(entity);
-        user.setTopos(listeTopos);
-        logger.info(" les topos de user: "+ user.getTopos());
-        userService.updateUser(user);
-
+        newTopo.setSites(entity.getSites());
         // enregistrement du topo dans chaque site concerné
         List<Site> sitesModifies=new ArrayList<Site>();
         for(int i=0;i<(newTopo.getSites()).size();i++){
@@ -113,9 +119,17 @@ public class TopoService {
         }
         newTopo.setSites(sitesModifies);
 
+        newTopo.setOwner(user);
+        // enregistrement du topo dans liste des topos de user
+        Collection<Topo> listeTopos = user.getTopos();
+        listeTopos.add(entity);
+        user.setTopos(listeTopos);
+        logger.info(" les topos de user: "+ user.getTopos());
+        userService.updateUser(user);
 
+        // enregistrement du topo dans la base de données
         entity = topoRepository.save(newTopo);
-        logger.info(" retour de l'entité de createOrUpdateTopo car l'Id n'existe pas");
+        logger.info(" retour de l'entité de createTopo car l'Id n'existe pas");
         return entity;
     }
 
@@ -133,19 +147,32 @@ public class TopoService {
                 Site site=listeSites.get(i);
                 site.setTopo(null);
             }
+
             // suppression de l'image du topo
-            imageService.deleteImageById(topoTrouve.getImage().getId());
+            if (topoTrouve.getImage()!=null) {
+                imageService.deleteImageById(topoTrouve.getImage().getId());
+            }
 
             // suppression de la réservation du topo
-            reservationTopoService.deleteReservationTopoById(topoTrouve.getReservation().getIdReservation());
+            if (topoTrouve.getReservation()!=null) {
+                User userTrouve = topoTrouve.getReservation().getUser();
+                Collection<Reservation> listeReservations = userTrouve.getReservations();
+                if (listeReservations != null) {
+                listeReservations.remove(topoTrouve.getReservation());
+                userTrouve.setReservations(listeReservations);
+                userService.updateUser(userTrouve);
+                }
+                reservationTopoService.deleteReservationTopoById(topoTrouve.getReservation().getIdReservation());
+            }
 
             // suppression des topos dans la liste du propriétaire
             List<Topo> listTopos=new ArrayList<Topo>();
-            listTopos.addAll(topoTrouve.getOwner().getTopos());
-            listTopos.remove(topoTrouve);
-            topoTrouve.getOwner().setTopos(listTopos);
-            userService.updateUser(topoTrouve.getOwner());
-
+            if(topoTrouve.getOwner().getTopos()!=null) {
+                listTopos.addAll(topoTrouve.getOwner().getTopos());
+                listTopos.remove(topoTrouve);
+                topoTrouve.getOwner().setTopos(listTopos);
+                userService.updateUser(topoTrouve.getOwner());
+            }
             // suppression du Topo de la base de données
             topoRepository.deleteById(topoTrouve.getIdTopo());
